Set things up for combat overhaul
