+++ mygame/commands/default_cmdsets.py
@@ -18,6 +18,7 @@ from evennia import default_cmds
 from evennia import CmdSet
 from commands import command
 from world import storynpcs
+from world.combat import combat
 
 class CharacterCmdSet(default_cmds.CharacterCmdSet):
     """
@@ -32,10 +33,10 @@ class CharacterCmdSet(default_cmds.CharacterCmdSet):
         """
         Populates the cmdset
         """
-        self.add(command.CmdAttack())
+        self.add(combat.CmdAttack())
         self.add(command.CmdCreateNPC())
         self.add(command.CmdEditNPC())
-        self.add(command.CmdEquip)
+        self.add(command.CmdEquip())
         self.add(storynpcs.CmdTalk())
         super().at_cmdset_creation()
         #
